+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.1.0
+- Fixes issues when passing requests through an unbounded throttle (`bytesPerSecond` = `Infinity`).
+- Adds public `onThroughputMetrics` metrics callback to the `BandwidthThrottleGroup`.
+
 # 1.0.1
 
 - Fixes issues on subsequent throttled requests in a group.